RoboJeff V1.2 Removed All previous functions and classes. Added Challenge Class for Challenge objects. Added Robot Class for the Robot's variables to be stored in. Added Motor Class for motor functions, currently missing and needing to be added.
This is a Theoretical pathing algorithm, independend of the monobrick reference library. Using Goniometry to decide how many degrees to turn and for how long to drive.
This is a work in progress and values need to be re-adjusted to actual measurements, these are mostly rough estimates.
This Code was first written in python as a draft and then converted to C#.

Todo
Add a way to see if a challenge is in between destination and origin.
Add functionality to Motor Class.
Add Sensory processing.